+++ crates/egui_inbox/src/lib.rs
@@ -6,8 +6,95 @@ use std::fmt::Debug;
 use std::mem;
 use std::sync::Arc;
 
-use egui::mutex::Mutex;
-use egui::{Context, Ui};
+use parking_lot::Mutex;
+
+/// Trait to request a repaint.
+pub trait RequestRepaintTrait {
+    /// Request a repaint.
+    fn request_repaint(&self);
+}
+
+impl<F> RequestRepaintTrait for F
+where
+    F: Fn() + Send + Sync + 'static,
+{
+    fn request_repaint(&self) {
+        self();
+    }
+}
+
+enum RequestRepaintInner {
+    #[cfg(feature = "egui")]
+    Ctx(egui::Context),
+    Box(Box<dyn RequestRepaintTrait + Send + Sync>),
+}
+
+/// Usually holds a reference to [egui::Context], but can also hold a boxed callback.
+#[derive(Debug)]
+pub struct RequestRepaintContext(RequestRepaintInner);
+
+impl RequestRepaintContext {
+    /// Create a new [RequestRepaintContext] from a callback function.
+    pub fn from_callback<F>(f: F) -> Self
+    where
+        F: Fn() + Send + Sync + 'static,
+    {
+        Self(RequestRepaintInner::Box(Box::new(f)))
+    }
+
+    /// Create a new [RequestRepaintContext] from something that implements [RequestRepaintTrait].
+    pub fn from_trait<T>(t: T) -> Self
+    where
+        T: RequestRepaintTrait + Send + Sync + 'static,
+    {
+        Self(RequestRepaintInner::Box(Box::new(t)))
+    }
+
+    /// Create a new [RequestRepaintContext] from an [egui::Context].
+    #[cfg(feature = "egui")]
+    pub fn from_egui_ctx(ctx: egui::Context) -> Self {
+        Self(RequestRepaintInner::Ctx(ctx))
+    }
+}
+
+impl RequestRepaintContext {
+    fn request_repaint(&self) {
+        match &self.0 {
+            RequestRepaintInner::Ctx(ctx) => ctx.request_repaint(),
+            RequestRepaintInner::Box(boxed) => boxed.request_repaint(),
+        }
+    }
+}
+
+impl Debug for RequestRepaintInner {
+    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
+        f.debug_struct("RequestRepaint").finish_non_exhaustive()
+    }
+}
+
+/// Trait to get a [RequestRepaintContext] from.
+pub trait AsRequestRepaint {
+    /// Should return a [RequestRepaintContext] that can be used to request a repaint.
+    fn as_request_repaint(&self) -> RequestRepaintContext;
+}
+
+#[cfg(feature = "egui")]
+mod egui_impl {
+    use crate::{AsRequestRepaint, RequestRepaintContext};
+    use egui::Context;
+
+    impl AsRequestRepaint for Context {
+        fn as_request_repaint(&self) -> RequestRepaintContext {
+            RequestRepaintContext::from_egui_ctx(self.clone())
+        }
+    }
+
+    impl AsRequestRepaint for egui::Ui {
+        fn as_request_repaint(&self) -> RequestRepaintContext {
+            RequestRepaintContext::from_egui_ctx(self.ctx().clone())
+        }
+    }
+}
 
 /// Utility to send messages to egui views from async functions, callbacks, etc. without
 /// having to use interior mutability.
@@ -42,9 +129,10 @@ use egui::{Context, Ui};
 ///     )
 /// }
 /// ```
-
 pub struct UiInbox<T> {
     state: Arc<Mutex<State<T>>>,
+    #[cfg(feature = "async")]
+    oneshot_channels: Vec<futures_channel::oneshot::Sender<()>>,
 }
 
 impl<T> Debug for UiInbox<T> {
@@ -55,13 +143,13 @@ impl<T> Debug for UiInbox<T> {
 
 #[derive(Debug)]
 struct State<T> {
-    ctx: Option<Context>,
+    ctx: Option<RequestRepaintContext>,
     queue: Vec<T>,
     dropped: bool,
 }
 
 impl<T> State<T> {
-    fn new(ctx: Option<Context>) -> Self {
+    fn new(ctx: Option<RequestRepaintContext>) -> Self {
         Self {
             ctx,
             queue: Vec::new(),
@@ -97,6 +185,13 @@ impl<T> Default for UiInbox<T> {
 
 impl<T> Drop for UiInbox<T> {
     fn drop(&mut self) {
+        #[cfg(feature = "async")]
+        {
+            self.oneshot_channels.drain(..).for_each(|tx| {
+                tx.send(()).ok();
+            });
+        }
+
         let mut state = self.state.lock();
         state.dropped = true;
     }
@@ -112,13 +207,17 @@ impl<T> UiInbox<T> {
     }
 
     /// Create a new inbox with a context.
-    pub fn new_with_ctx(ctx: Context) -> Self {
-        Self::_new(Some(ctx))
+    pub fn new_with_ctx(ctx: &impl AsRequestRepaint) -> Self {
+        Self::_new(Some(ctx.as_request_repaint()))
     }
 
-    fn _new(ctx: Option<Context>) -> Self {
+    fn _new(ctx: Option<RequestRepaintContext>) -> Self {
         let state = Arc::new(Mutex::new(State::new(ctx)));
-        Self { state }
+        Self {
+            state,
+            #[cfg(feature = "async")]
+            oneshot_channels: Vec::new(),
+        }
     }
 
     /// Create a inbox and a sender for it.
@@ -129,7 +228,7 @@ impl<T> UiInbox<T> {
     }
 
     /// Create a inbox with a context and a sender for it.
-    pub fn channel_with_ctx(ctx: Context) -> (UiInboxSender<T>, Self) {
+    pub fn channel_with_ctx(ctx: &impl AsRequestRepaint) -> (UiInboxSender<T>, Self) {
         let inbox = Self::new_with_ctx(ctx);
         let sender = inbox.sender();
         (sender, inbox)
@@ -137,8 +236,8 @@ impl<T> UiInbox<T> {
 
     /// Set the [Context] to use for requesting repaints.
     /// Usually this is not needed, since the [Context] is grabbed from the [Ui] passed to [UiInbox::read].
-    pub fn set_ctx(&mut self, ctx: Context) {
-        self.state.lock().ctx = Some(ctx);
+    pub fn set_ctx(&mut self, ctx: &impl AsRequestRepaint) {
+        self.state.lock().ctx = Some(ctx.as_request_repaint());
     }
 
     /// Returns an iterator over all items sent to the inbox.
@@ -147,10 +246,10 @@ impl<T> UiInbox<T> {
     /// The ui is only passed here so we can grab a reference to [Context].
     /// This is mostly done for convenience, so you don't have to pass a reference to [Context]
     /// to every struct that uses an inbox on creation.
-    pub fn read(&self, ui: &mut Ui) -> impl Iterator<Item = T> {
+    pub fn read(&self, ui: &impl AsRequestRepaint) -> impl Iterator<Item = T> {
         let mut state = self.state.lock();
         if state.ctx.is_none() {
-            state.ctx = Some(ui.ctx().clone());
+            state.ctx = Some(ui.as_request_repaint());
         }
         mem::take(&mut state.queue).into_iter()
     }
@@ -158,7 +257,7 @@ impl<T> UiInbox<T> {
     /// Same as [UiInbox::read], but you don't need to pass a reference to [Ui].
     /// If you use this, make sure you set the [Context] with [UiInbox::set_ctx] or
     /// [UiInbox::new_with_ctx] manually.
-    pub fn read_without_ui(&self) -> impl Iterator<Item = T> {
+    pub fn read_without_ctx(&self) -> impl Iterator<Item = T> {
         let mut state = self.state.lock();
         mem::take(&mut state.queue).into_iter()
     }
@@ -171,10 +270,10 @@ impl<T> UiInbox<T> {
     /// The ui is only passed here so we can grab a reference to [Context].
     /// This is mostly done for convenience, so you don't have to pass a reference to [Context]
     /// to every struct that uses an inbox on creation.
-    pub fn replace(&self, ui: &mut Ui, target: &mut T) -> bool {
+    pub fn replace(&self, ui: &impl AsRequestRepaint, target: &mut T) -> bool {
         let mut state = self.state.lock();
         if state.ctx.is_none() {
-            state.ctx = Some(ui.ctx().clone());
+            state.ctx = Some(ui.as_request_repaint());
         }
 
         let item = mem::take(&mut state.queue).pop();
@@ -189,7 +288,7 @@ impl<T> UiInbox<T> {
     /// Same as [UiInbox::replace], but you don't need to pass a reference to [Ui].
     /// If you use this, make sure you set the [Context] with [UiInbox::set_ctx] or
     /// [UiInbox::new_with_ctx] manually.
-    pub fn replace_without_ui(&self, target: &mut T) -> bool {
+    pub fn replace_without_ctx(&self, target: &mut T) -> bool {
         let mut state = self.state.lock();
         let item = mem::take(&mut state.queue).pop();
         if let Some(item) = item {
@@ -204,6 +303,95 @@ impl<T> UiInbox<T> {
     pub fn sender(&self) -> UiInboxSender<T> {
         UiInboxSender {
             state: self.state.clone(),
+        }
+    }
+}
+
+#[cfg(feature = "async")]
+mod async_impl {
+    use std::pin::{pin, Pin};
+    use std::task::{Context, Poll};
+
+    use futures::{select, FutureExt, Sink, SinkExt, StreamExt};
+
+    use hello_egui_utils::{spawn, MaybeSend};
+
+    use crate::{SendError, UiInbox, UiInboxSender};
+
+    impl<T> UiInbox<T> {
+        /// Spawns a future that will automatically be cancelled when the inbox is dropped.
+        /// Make sure your future is safe to cancel (It may stop at any await point).
+        ///
+        /// If you want to spawn a future that should definitely run to completion, use [UiInbox::spawn_detached] instead.
+        pub fn spawn<F>(&mut self, f: impl FnOnce(UiInboxSender<T>) -> F)
+        where
+            F: std::future::Future<Output = ()> + MaybeSend + 'static,
+        {
+            let (tx, mut rx) = futures_channel::oneshot::channel();
+            self.oneshot_channels.push(tx);
+
+            let sender = self.sender();
+            let future = f(sender).fuse();
+
+            spawn(async move {
+                let mut future = pin!(future);
+
+                select! {
+                    _ = future => {},
+                    _ = rx => {},
+                }
+            });
+        }
+
+        /// Spawns a future that will **not** be cancelled when the inbox is dropped.
+        pub fn spawn_detached<F>(&mut self, f: impl FnOnce(UiInboxSender<T>) -> F)
+        where
+            F: std::future::Future<Output = ()> + Send + 'static,
+        {
+            let sender = self.sender();
+            let future = f(sender);
+            spawn(future);
+        }
+    }
+
+    impl<T> UiInboxSender<T> {
+        /// Send each item of a stream to the inbox, as they come in.
+        pub async fn send_stream(
+            &mut self,
+            stream: impl futures::Stream<Item = T> + Send + 'static,
+        ) -> Result<(), SendError<T>> {
+            let stream = stream.map(|i| Ok(i));
+            let mut stream = pin!(stream);
+            self.send_all(&mut stream).await
+        }
+    }
+
+    impl<T> Sink<T> for UiInboxSender<T> {
+        type Error = SendError<T>;
+
+        fn poll_ready(
+            self: Pin<&mut Self>,
+            _cx: &mut Context<'_>,
+        ) -> Poll<Result<(), Self::Error>> {
+            Poll::Ready(Ok(()))
+        }
+
+        fn start_send(self: Pin<&mut Self>, item: T) -> Result<(), Self::Error> {
+            UiInboxSender::send(&self, item)
+        }
+
+        fn poll_flush(
+            self: Pin<&mut Self>,
+            _cx: &mut Context<'_>,
+        ) -> Poll<Result<(), Self::Error>> {
+            Poll::Ready(Ok(()))
+        }
+
+        fn poll_close(
+            self: Pin<&mut Self>,
+            _cx: &mut Context<'_>,
+        ) -> Poll<Result<(), Self::Error>> {
+            Poll::Ready(Ok(()))
         }
     }
 }