+++ fancy-example/src/main.rs
@@ -73,8 +73,8 @@ impl App {
 
 impl eframe::App for App {
     fn update(&mut self, ctx: &Context, _frame: &mut Frame) {
-        self.inbox.set_ctx(ctx.clone());
-        self.inbox.read_without_ui().for_each(|msg| match msg {
+        self.inbox.set_ctx(ctx);
+        self.inbox.read_without_ctx().for_each(|msg| match msg {
             FancyMessage::SelectPage(active) => {
                 self.sidebar.active = active;
             }